update: 模型管理目录结构调整
