+++ src/ui/src/components/model-manage/_choose-icon.vue
