+++ src/ui/src/views/model-association/_detail.vue
