+++ src/ui/src/views/model-manage/children/index.vue
@@ -84,7 +84,7 @@
     import thePropertyGroup from './_property-group.vue'
     import theField from './field'
     import theRelation from './relation'
-    import theChooseIcon from '../_choose-icon'
+    import theChooseIcon from '@/components/model-manage/_choose-icon'
     import { mapActions, mapGetters, mapMutations } from 'vuex'
     export default {
         components: {
