+++ src/ui/src/views/model-topology/children/display-config.vue
