+++ src/ui/src/views/model-topology/children/relation-detail.vue
